Update TCP logic

+++ src/ScriptHubMain.py
@@ -1,9 +1,16 @@
 #!/usr/bin/python
 import regist, framerun_base, time, sig_handle, traceback, log, logging
 from signal import SIGTERM
+from signal import SIGKILL
 import sys, os, atexit
 
+pid_file = "daemon.pid"
+
 class Daemon:
+    """
+    A generic daemon class.
+    Usage: subclass the Daemon class and override the run() method
+    """
     def __init__(self, pidfile, stdin='/dev/null', stdout='/dev/null', stderr='/dev/null'):
         self.stdin = stdin
         self.stdout = stdout
@@ -11,26 +18,36 @@ class Daemon:
         self.pidfile = pidfile
 
     def daemonize(self):
+        """
+        do the UNIX double-fork magic, see Stevens' "Advanced
+        Programming in the UNIX Environment" for details (ISBN 0201563177)
+        http://www.erlenstar.demon.co.uk/unix/faq_2.html#SEC16
+        """
         try:
             pid = os.fork()
             if pid > 0:
                 # exit first parent
                 sys.exit(0)
         except OSError, e:
-            sys.stderr.write("fork error: %d (%s)\n" % (e.errno, e.strerror))
+            sys.stderr.write("fork #1 failed: %d (%s)\n" % (e.errno, e.strerror))
             sys.exit(1)
 
+        # decouple from parent environment
+#        os.chdir("/")
         os.setsid()
         os.umask(0)
 
+        # do second fork
         try:
             pid = os.fork()
             if pid > 0:
+                # exit from second parent
                 sys.exit(0)
         except OSError, e:
-            sys.stderr.write("fork error: %d (%s)\n" % (e.errno, e.strerror))
+            sys.stderr.write("fork #2 failed: %d (%s)\n" % (e.errno, e.strerror))
             sys.exit(1)
 
+        # redirect standard file descriptors
         sys.stdout.flush()
         sys.stderr.flush()
         si = file(self.stdin, 'r')
@@ -40,6 +57,7 @@ class Daemon:
         os.dup2(so.fileno(), sys.stdout.fileno())
         os.dup2(se.fileno(), sys.stderr.fileno())
 
+        # write pidfile
         atexit.register(self.delpid)
         pid = str(os.getpid())
         file(self.pidfile,'w+').write("%s\n" % pid)
@@ -51,6 +69,7 @@ class Daemon:
         """
         Start the daemon
         """
+        # Check for a pidfile to see if the daemon already runs
         try:
             pf = file(self.pidfile,'r')
             pid = int(pf.read().strip())
@@ -59,7 +78,7 @@ class Daemon:
             pid = None
 
         if pid:
-            message = "pidfile %s already exist. Daemon already running...\n"
+            message = "pidfile %s already exist. Daemon already running?\n"
             sys.stderr.write(message % self.pidfile)
             sys.exit(1)
 
@@ -98,7 +117,7 @@ class Daemon:
             pid = None
 
         if not pid:
-            message = "pidfile %s does not exist. Daemon not running...\n"
+            message = "pidfile %s does not exist. Daemon not running?\n"
             sys.stderr.write(message % self.pidfile)
             return # not an error in a restart
 
@@ -106,9 +125,18 @@ class Daemon:
         try:
             while 1:
                 os.kill(pid, SIGTERM)
-                time.sleep(0.1)
+                time.sleep(1)
+                count = 0
+                sys.stderr.write("Waiting ScriptHubMaster(%d) clean environment(wait 120s)...\n" % pid)
+                while os.path.exists(self.pidfile):
+                    time.sleep(1)
+                    count += 1
+                    if count > 120:
+                        sys.stderr.write("ScriptHubMaster(%d) still running, send SIGKILL\n" % pid)
+                        os.kill(pid, SIGKILL)
+                        os.remove(self.pidfile)
+                        sys.exit(0)
                 sys.stderr.write("ScriptHubMaster(%d) killing success\n" % pid)
-                os.remove(self.pidfile)
                 sys.exit(0)
         except OSError, err:
             err = str(err)
@@ -137,9 +165,9 @@ class Daemon:
             pid = None
 
         if not pid:
-            message = "pidfile %s does not exist. Daemon not running...\n"
+            message = "pidfile %s does not exist. Daemon not running?\n"
             sys.stderr.write(message % self.pidfile)
-            return 
+            return
 
         # Try send signal to the daemon process
         try:
@@ -166,7 +194,7 @@ class Daemon:
             pid = None
 
         if not pid:
-            message = "pidfile %s does not exist. Daemon not running...\n"
+            message = "pidfile %s does not exist. Daemon not running?\n"
             sys.stderr.write(message % self.pidfile)
             return # not an error in a restart
 
@@ -223,6 +251,8 @@ class ScriptHubDaemon(Daemon):
             sig_handle.signal_regist()
 
             # Init Time & Tcp module
+            regist.time_module_conf_init()
+            regist.tcp_module_conf_init()
             time_list = regist.time_module_init()
             tcp_list = regist.tcp_module_init()
 
@@ -237,10 +267,20 @@ class ScriptHubDaemon(Daemon):
                 tcp_global.epoll_add(i)
 
             while True:
+                if sig_handle.Stop_Flag == 1:
+                    logging.info("Master process start to clean environment")
+                    time_global.process_exit()
+                    os.remove(pid_file)
+                    logging.info("Master process environment clean finish and exit")
+                    sys.exit(0)
+
+                # Event Runnning Loop
                 if tcp_global.event_count() > 0:
                     tcp_global.epoll_loop()
                 if time_global.event_count() > 0:
                     time_global.time_loop()
+
+
                 if sig_handle.Restart_Flag == 1:
                     # Restart Python Program
                     logging.info("Start to Restart Program")
@@ -262,14 +302,16 @@ class ScriptHubDaemon(Daemon):
                             time_global.signal_loop(i)
                             sig_handle.User_signal.remove(i)
 
-                time.sleep(1)
+                time.sleep(0.01)
 
         except Exception, e:
+            mailname = "EventModuleMaster Down"
             error_buff = str(e) + ":" + traceback.format_exc()
             logging.error(error_buff)
+            os.remove(pid_file)
 
 if __name__ == '__main__':
-    daemon = ScriptHubDaemon("daemon.pid")
+    daemon = ScriptHubDaemon(pid_file)
     if len(sys.argv) >= 2:
         if 'start' == sys.argv[1]:
             daemon.start()
@@ -283,11 +325,11 @@ if __name__ == '__main__':
             if len(sys.argv) < 3:
                 print "usage: %s signal [signal_num]" % sys.argv[0]
                 sys.exit(2)
-            if len(sys.argv) >= 4:
+            if len(sys.argv) == 4:
                 signal_num = int(sys.argv[2])
                 print "Get specify signal with string (%s)" % sys.argv[3]
                 daemon.signal_specify(signal_num, sys.argv[3])
-            else:
+            elif len(sys.argv) == 3:
                 signal_num = int(sys.argv[2])
                 daemon.signal(signal_num)
         else:
@@ -295,5 +337,5 @@ if __name__ == '__main__':
             sys.exit(2)
         sys.exit(0)
     else:
-        print "usage: %s start|stop|status|signal [signal_num]|signal [signal_num] [module_name]" % sys.argv[0]
+        print "usage: %s start|stop|restart|status|signal [signal_num]|signal [signal_num] [module_name]" % sys.argv[0]
         sys.exit(2)
